# resources/contoso-store-api-swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "ContosoStore.API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://aiagentwks-contoso-store-mbo43n.azurewebsites.net"
    }
  ],
  "paths": {
    "/api/bikes": {
      "get": {
        "operationId": "getBikes",
        "tags": [
          "Store"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/bikes/{id}": {
      "get": {
        "operationId": "getBikeById",
        "tags": [
          "Store"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/orders": {
      "post": {
        "operationId": "createOrder",
        "tags": [
          "Store"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BikeOrderRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BikeOrderRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BikeOrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/orders/{orderId}": {
      "get": {
        "operationId": "getOrderById",
        "tags": [
          "Store"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BikeOrderRequest": {
        "type": "object",
        "properties": {
          "bikeId": {
            "type": "integer",
            "format": "int32"
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "customerEmail": {
            "type": "string",
            "nullable": true
          },
          "shippingAddress": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}